chore: make v1beta1 strucs more readable
