+++ apis/v1beta1/targetallocator_types.go
@@ -20,6 +20,42 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
+func init() {
+	SchemeBuilder.Register(&TargetAllocator{}, &TargetAllocatorList{})
+}
+
+//+kubebuilder:object:root=true
+//+kubebuilder:subresource:status
+
+// TargetAllocator is the Schema for the targetallocators API.
+type TargetAllocator struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+
+	Spec   TargetAllocatorSpec   `json:"spec,omitempty"`
+	Status TargetAllocatorStatus `json:"status,omitempty"`
+}
+
+//+kubebuilder:object:root=true
+
+// TargetAllocatorList contains a list of TargetAllocator.
+type TargetAllocatorList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata,omitempty"`
+	Items           []OpenTelemetryCollector `json:"items"`
+}
+
+// TargetAllocatorStatus defines the observed state of Target Allocator.
+type TargetAllocatorStatus struct {
+	// Version of the managed Target Allocator (operand)
+	// +optional
+	Version string `json:"version,omitempty"`
+
+	// Image indicates the container image to use for the Target Allocator.
+	// +optional
+	Image string `json:"image,omitempty"`
+}
+
 // TargetAllocatorSpec defines the desired state of TargetAllocator.
 type TargetAllocatorSpec struct {
 	// Common defines fields that are common to all OpenTelemetry CRD workloads.
@@ -70,40 +106,4 @@ type TargetAllocatorPrometheusCR struct {
 	// ServiceMonitor's meta labels. The requirements are ANDed.
 	// +optional
 	ServiceMonitorSelector *metav1.LabelSelector `json:"serviceMonitorSelector,omitempty"`
-}
-
-// TargetAllocatorStatus defines the observed state of Target Allocator.
-type TargetAllocatorStatus struct {
-	// Version of the managed Target Allocator (operand)
-	// +optional
-	Version string `json:"version,omitempty"`
-
-	// Image indicates the container image to use for the Target Allocator.
-	// +optional
-	Image string `json:"image,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
-
-// TargetAllocator is the Schema for the targetallocators API.
-type TargetAllocator struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Spec   TargetAllocatorSpec   `json:"spec,omitempty"`
-	Status TargetAllocatorStatus `json:"status,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-
-// TargetAllocatorList contains a list of TargetAllocator.
-type TargetAllocatorList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []OpenTelemetryCollector `json:"items"`
-}
-
-func init() {
-	SchemeBuilder.Register(&TargetAllocator{}, &TargetAllocatorList{})
 }
